got rid of weird search ui
