+++ frontend/app/(tabs)/search.tsx
@@ -15,17 +15,14 @@ import {
   View,
 } from "react-native";
 
-// ---- Firestore (v9 modular) ----
+// Import search service
 import {
-  collection,
-  endAt,
-  getDocs,
-  orderBy,
-  query as q,
-  limit as qLimit,
-  startAt,
-} from "firebase/firestore";
-import { db } from "../../config/firebase"; // <-- your initialized Firestore instance
+  normalize,
+  ResultItem,
+  searchFirestore,
+  TabKey,
+  useDebounced,
+} from "../../components/searchService";
 
 const COLORS = {
   text: "#111827",
@@ -35,41 +32,6 @@ const COLORS = {
   accent: "#6FA076",
 };
 const BG = "#FFFAF0";
-
-type TabKey = "locations" | "members";
-
-type Spot = {
-  id: string;
-  name: string;
-  name_lower?: string;
-  city?: string;
-  category?: string;
-};
-type UserLite = {
-  id: string;
-  handle?: string;
-  username_lower?: string;
-  displayName?: string;
-  rank?: number;
-  followers?: number;
-};
-
-type ResultItem = {
-  id: string;
-  title: string;
-  subtitle?: string;
-  spotData?: SpotCardType;
-};
-
-function toTitleCase(str: string) {
-  return str
-    .toLowerCase()
-    .split(" ")
-    .filter(Boolean) // remove extra spaces
-    .map(w => w[0].toUpperCase() + w.slice(1))
-    .join(" ");
-}
-
 
 export default function SearchScreen() {
   const [query, setQuery] = useState("");
@@ -83,11 +45,7 @@ export default function SearchScreen() {
   const [error, setError] = useState<string | null>(null);
 
   // Recents (very simple in-memory; persist to AsyncStorage if you like)
-  const [recents, setRecents] = useState<ResultItem[]>([
-    { id: "u1", title: "Will Levy", subtitle: "@wlevy" },
-    { id: "u2", title: "Bell Tran", subtitle: "@belltran" },
-    { id: "u3", title: "Jane Doe", subtitle: "@janedoe" },
-  ]);
+  const [recents, setRecents] = useState<ResultItem[]>([]);
 
   // Simple result cache: key = `${tab}|${q}`
   const cacheRef = useRef<Record<string, ResultItem[]>>({});
@@ -165,24 +123,28 @@ export default function SearchScreen() {
   const handleSpotPress = (spot: SpotCardType) => {
     setOverlayOpen(false);
     router.push({
-      pathname: '/spot-detail',
+      pathname: "/spot-detail",
       params: {
-        spotData: JSON.stringify(spot)
-      }
+        spotData: JSON.stringify(spot),
+      },
     });
   };
 
   return (
     <SafeAreaView style={styles.safe}>
       <ThemedView style={styles.container}>
-        <ThemedText type="title" style={styles.header}>Search</ThemedText>
+        <ThemedText type="title" style={styles.header}>
+          Search
+        </ThemedText>
 
         {/* Clean search bar */}
         <View style={styles.searchBarContainer}>
-          <View style={[
-            styles.searchInputContainer,
-            isFocused && styles.searchInputFocused
-          ]}>
+          <View
+            style={[
+              styles.searchInputContainer,
+              isFocused && styles.searchInputFocused,
+            ]}
+          >
             <Text style={styles.searchIcon}>🔍</Text>
             <TextInput
               value={query}
@@ -203,25 +165,26 @@ export default function SearchScreen() {
         {/* Clean description */}
         <View style={styles.descriptionContainer}>
           <ThemedText style={styles.descriptionText}>
-            Find places, connect with people, and discover content from our community.
+            Find places, connect with people, and discover content from our
+            community.
           </ThemedText>
-          
+
           {/* Subtle quick actions */}
           <View style={styles.quickActionsContainer}>
             <View style={styles.quickActionsRow}>
-              <Pressable 
+              <Pressable
                 style={styles.quickActionPill}
                 onPress={() => setQuery("coffee")}
               >
                 <ThemedText style={styles.quickActionText}>Coffee</ThemedText>
               </Pressable>
-              <Pressable 
+              <Pressable
                 style={styles.quickActionPill}
                 onPress={() => setQuery("hiking")}
               >
                 <ThemedText style={styles.quickActionText}>Hiking</ThemedText>
               </Pressable>
-              <Pressable 
+              <Pressable
                 style={styles.quickActionPill}
                 onPress={() => setQuery("restaurants")}
               >
@@ -337,51 +300,8 @@ export default function SearchScreen() {
                         </Pressable>
                       ))}
                     </ScrollView>
+                   
 
-                    <View style={styles.sectionHeaderRow}>
-                      <Text style={styles.sectionTitle}>Suggested for you</Text>
-                      <Pressable>
-                        <Text style={styles.link}>See all</Text>
-                      </Pressable>
-                    </View>
-
-                    {tab === "locations" ? (
-                      <>
-                        <ResultRow
-                          title="Cafe Lumen"
-                          subtitle="Cambridge • Coffee"
-                          onPress={onSelectResult}
-                        />
-                        <ResultRow
-                          title="Shiso Kitchen"
-                          subtitle="Somerville • Japanese"
-                          onPress={onSelectResult}
-                        />
-                        <ResultRow
-                          title="Riverview Diner"
-                          subtitle="Boston • American"
-                          onPress={onSelectResult}
-                        />
-                      </>
-                    ) : (
-                      <>
-                        <ResultRow
-                          title="@charlie"
-                          subtitle="Rank #12 • 240 followers"
-                          onPress={onSelectResult}
-                        />
-                        <ResultRow
-                          title="@mia"
-                          subtitle="Rank #33 • 120 followers"
-                          onPress={onSelectResult}
-                        />
-                        <ResultRow
-                          title="@alex"
-                          subtitle="Rank #58 • 90 followers"
-                          onPress={onSelectResult}
-                        />
-                      </>
-                    )}
                   </>
                 ) : (
                   <>
@@ -432,127 +352,6 @@ export default function SearchScreen() {
       </ThemedView>
     </SafeAreaView>
   );
-}
-
-function useDebounced<T>(value: T, ms: number) {
-  const [v, setV] = useState(value);
-  useEffect(() => {
-    const t = setTimeout(() => setV(value), ms);
-    return () => clearTimeout(t);
-  }, [value, ms]);
-  return v;
-}
-
-// Strip diacritics + lowercase to match your `*_lower` fields
-function normalize(s: string) {
-  return (
-    s
-      .normalize("NFKD")
-      // @ts-ignore — Unicode property escapes supported in Hermes/JSI RN 0.72+
-      .replace(/\p{Diacritic}/gu, "")
-      .toLowerCase()
-      .trim()
-  );
-}
-
-/**
- * Firestore prefix search
- * - locations -> collection("spots"), orderBy("name_lower")
- * - members   -> collection("users"), orderBy("username_lower")
- * Returns a unified ResultItem[]
- *
- * NOTE: Ensure you have composite indexes if you add extra where() later.
- */
-async function searchFirestore(tab: TabKey, queryInput: string): Promise<ResultItem[]> {
-  if (!queryInput) return [];
-
-  console.log("Searching for:", queryInput);
-
-  if (tab === "locations") {
-    const ref = collection(db, "spots");
-    const formattedQuery = toTitleCase(queryInput);
-    
-    const queryRef = q(
-      ref,
-      orderBy("name"),
-      startAt(formattedQuery),
-      endAt(formattedQuery + "\uf8ff"),
-      qLimit(20)
-    );
-    const snap = await getDocs(queryRef);
-    // console.log("Found documents:", snap.docs.length);
-    return snap.docs.map(d => {
-      const data = d.data() as SpotCardType;
-      // console.log("Document data:", data);
-      // Ensure the spot data includes the document ID
-      const spotDataWithId = {
-        ...data,
-        id: d.id || `spot_${Date.now()}_${Math.random()}`
-      };
-      return { 
-        id: d.id || `spot_${Date.now()}_${Math.random()}`, 
-        title: data?.name || "(untitled)", 
-        subtitle: buildLocationSubtitle(data),
-        spotData: spotDataWithId // Add the full spot data with ID
-      };
-    });
-  } else {
-    const ref = collection(db, "users");
-    const formattedQuery = toTitleCase(queryInput);
-    
-    const queryRef = q(
-      ref,
-      orderBy("displayName"),
-      startAt(formattedQuery),
-      endAt(formattedQuery + "\uf8ff"),
-      qLimit(20)
-    );
-    const snap = await getDocs(queryRef);
-    console.log("Found users:", snap.docs.length);
-    return snap.docs.map(d => {
-      const u = d.data() as UserLite;
-      const title = u?.handle ? `@${u.handle}` : u?.displayName || "(user)";
-      const subtitle =
-        u?.displayName && u?.handle
-          ? `${u.displayName}`
-          : u?.followers || u?.rank
-          ? `Rank #${u.rank ?? "-"} • ${u.followers ?? 0} followers`
-          : undefined;
-      return { 
-        id: d.id || `user_${Date.now()}_${Math.random()}`, 
-        title, 
-        subtitle 
-      };
-    });
-  }
-}
-
-function buildLocationSubtitle(s: SpotCardType) {
-  if (!s) return undefined;
-  const bits = [];
-  
-  // Add location
-  if (s.location?.address) {
-    bits.push(s.location.address);
-  }
-  
-  // Add category
-  if (s.category) {
-    const categoryFormatted = s.category.replace(/_/g, ' ').toUpperCase();
-    bits.push(categoryFormatted);
-  }
-  
-  // Add rating if available
-  if (s.averageRating > 0) {
-    bits.push(`⭐ ${s.averageRating.toFixed(1)}`);
-  }
-  
-  // Add verification badge
-  if (s.isVerified) {
-    bits.push('✓ Verified');
-  }
-  
-  return bits.length ? bits.join(" • ") : undefined;
 }
 
 /* ---------- Presentational ---------- */
@@ -623,12 +422,12 @@ const styles = StyleSheet.create({
     textTransform: "lowercase",
   },
 
-  searchBarContainer: { 
+  searchBarContainer: {
     marginBottom: 20,
   },
   searchInputContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
+    flexDirection: "row",
+    alignItems: "center",
     backgroundColor: COLORS.bg,
     borderRadius: 12,
     borderWidth: 1,
@@ -664,20 +463,20 @@ const styles = StyleSheet.create({
     marginTop: 4,
   },
   quickActionsRow: {
-    flexDirection: 'row',
+    flexDirection: "row",
     gap: 12,
   },
   quickActionPill: {
-    backgroundColor: '#F8F9FA',
+    backgroundColor: "#F8F9FA",
     borderRadius: 16,
     paddingHorizontal: 12,
     paddingVertical: 6,
     borderWidth: 1,
-    borderColor: '#E9ECEF',
+    borderColor: "#E9ECEF",
   },
   quickActionText: {
     fontSize: 14,
-    fontWeight: '500',
+    fontWeight: "500",
     color: COLORS.accent,
   },
